refactor: extract module support into separate Modulizer class
