+++ src/incorporate.ts
@@ -1,17 +1,19 @@
 import {createElement, forwardRef} from 'react';
 import {Scope} from './scope';
 import {ScopeContext} from './context';
-import Incorporator from './Incorporator';
+import {Modulizer} from './Modulizer'
+
+
 
 const wrapperComponents: Map<any, React.ComponentType<any>> = new Map();
-
+const identity = x => x
 export function incorporate(type: any) {
   if (!wrapperComponents.has(type)) {
     wrapperComponents.set(
       type,
       forwardRef<any, any>((props, ref) =>
         createElement(ScopeContext.Consumer, null, (scope: Scope) =>
-          createElement(Incorporator, {
+          createElement(Modulizer, {
             targetProps: props,
             targetRef: ref,
             target: type,
