+++ src/index.ts
@@ -4,5 +4,5 @@ export {Scope} from './scope';
 export {ReactSource} from './ReactSource';
 export {h} from './h';
 export {incorporate} from './incorporate';
-export {setModules} from './Incorporator'
+export {setModules} from './Modulizer'
 export {StreamRenderer} from './StreamRenderer';
